Clean up examples to generic editor
Just put them on the regular samples bundles, as there is no benefit in
having them separated.
