+++ eclipse/fr.opensagres.language.textmate.eclipse.samples/plugin.xml
@@ -35,4 +35,20 @@
       </editor>
    </extension>
    
+   <extension
+         point="org.eclipse.ui.genericeditor.presentationReconcilers">
+      <presentationReconciler
+            class="fr.opensagres.language.textmate.eclipse.samples.csharp.CSharpPresentationReconclier"
+            contentType="org.eclipse.languageserver.languages.csharp">
+      </presentationReconciler>
+      <presentationReconciler
+            class="fr.opensagres.language.textmate.eclipse.samples.json.JSONPresentationReconcilier"
+            contentType="org.eclipse.languageserver.languages.json">
+      </presentationReconciler>
+      <presentationReconciler
+            class="fr.opensagres.language.textmate.eclipse.samples.css.CSSPresentationReconcilier"
+            contentType="org.eclipse.languageserver.languages.css">
+      </presentationReconciler>      
+   </extension>
+   
 </plugin>
