+++ eclipse/fr.opensagres.language.textmate.eclipse.samples/src/main/java/fr/opensagres/language/textmate/eclipse/samples/csharp/style.css
