--- eclipse/org.eclipse.languageserver.languages.textmate/META-INF/MANIFEST.MF
@@ -1,13 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: Support for Languages using Language Servers (C#, CSS, LESS, SCSS, JSON)
-Bundle-SymbolicName: org.eclipse.languageserver.languages.textmate;singleton:=true
-Bundle-Version: 0.1.0.qualifier
-Bundle-Vendor: Eclipse.org
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Require-Bundle: org.eclipse.core.runtime,
- org.eclipse.ui,
- org.eclipse.jface.text,
- org.eclipse.ui.editors;bundle-version="3.10.0",
- fr.opensagres.language.textmate.core;bundle-version="1.0.0",
- fr.opensagres.language.textmate.eclipse;bundle-version="1.0.0"
