--- eclipse/org.eclipse.languageserver.languages.textmate/plugin.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<?eclipse version="3.4"?>
-<plugin>
-   <extension
-         point="org.eclipse.ui.genericeditor.presentationReconcilers">
-      <presentationReconciler
-            class="org.eclipse.languageserver.languages.csharp.CSharpPresentationReconclier"
-            contentType="org.eclipse.languageserver.languages.csharp">
-      </presentationReconciler>
-      <presentationReconciler
-            class="org.eclipse.languageserver.languages.json.JSONPresentationReconcilier"
-            contentType="org.eclipse.languageserver.languages.json">
-      </presentationReconciler>
-      <presentationReconciler
-            class="org.eclipse.languageserver.languages.css.CSSPresentationReconcilier"
-            contentType="org.eclipse.languageserver.languages.css">
-      </presentationReconciler>      
-   </extension>
-</plugin>
