--- eclipse/org.eclipse.languageserver.languages.textmate/pom.xml
@@ -1,37 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <artifactId>org.eclipse.languageserver.languages</artifactId>
-	<parent>
-  	<groupId>org.eclipse.languageserver</groupId>
-		<artifactId>parent</artifactId>
-	  <version>0.1.0-SNAPSHOT</version>
-	</parent>
-  <packaging>eclipse-plugin</packaging>
-  
-  <build>
-  	<plugins>
-  		<plugin>
-  			<groupId>org.eclipse.tycho</groupId>
-  			<artifactId>target-platform-configuration</artifactId>
-  			<version>0.25.0</version>
-  			<configuration>
-  				<pomDependencies>consider</pomDependencies>
-  			</configuration>
-  		</plugin>
-  	</plugins>
-  </build>
-  
-  <dependencies>
-  	<dependency>
-  		<groupId>com.google.code.gson</groupId>
-  		<artifactId>gson</artifactId>
-  		<version>2.5</version>
-  	</dependency>
-		<dependency>
-			<groupId>org.apache.servicemix.bundles</groupId>
-			<artifactId>org.apache.servicemix.bundles.jsr305</artifactId>
-			<version>3.0.1_1</version>
-		</dependency>
-  </dependencies>
-  
-</project>
