--- eclipse/org.eclipse.languageserver.languages.textmate/src/org/eclipse/languageserver/languages/json/style_Solarized-light.css
@@ -1,41 +1,0 @@
-.comment { color: rgba(147, 161, 161, 1); }
-.string { color: rgba(42, 161, 152, 1); }
-.string.regexp { color: rgba(211, 1, 2, 1); }
-.constant.numeric { color: rgba(211, 54, 130, 1); }
-.variable.language { color: rgba(38, 139, 210, 1); }
-.variable.other { color: rgba(38, 139, 210, 1); }
-.keyword { color: rgba(133, 153, 0, 1); }
-.storage { font-weight: bold; color: rgba(7, 54, 66, 1); }
-.entity.name.class { color: rgba(38, 139, 210, 1); }
-.entity.name.type.class { color: rgba(38, 139, 210, 1); }
-.entity.name.function { color: rgba(38, 139, 210, 1); }
-.punctuation.definition.variable { color: rgba(133, 153, 0, 1); }
-.punctuation.section.embedded.begin { color: rgba(211, 1, 2, 1); }
-.punctuation.section.embedded.end { color: rgba(211, 1, 2, 1); }
-.constant.language { color: rgba(181, 137, 0, 1); }
-.meta.preprocessor { color: rgba(181, 137, 0, 1); }
-.support.function.construct { color: rgba(211, 1, 2, 1); }
-.keyword.other.new { color: rgba(211, 1, 2, 1); }
-.constant.character { color: rgba(203, 75, 22, 1); }
-.constant.other { color: rgba(203, 75, 22, 1); }
-.entity.other.inherited-class {  }
-.variable.parameter {  }
-.entity.name.tag { color: rgba(38, 139, 210, 1); }
-.punctuation.definition.tag.begin { color: rgba(147, 161, 161, 1); }
-.punctuation.definition.tag.end { color: rgba(147, 161, 161, 1); }
-.entity.other.attribute-name { color: rgba(147, 161, 161, 1); }
-.support.function { color: rgba(38, 139, 210, 1); }
-.punctuation.separator.continuation { color: rgba(211, 1, 2, 1); }
-.support.constant {  }
-.support.type { color: rgba(133, 153, 0, 1); }
-.support.class { color: rgba(133, 153, 0, 1); }
-.support.type.exception { color: rgba(203, 75, 22, 1); }
-.support.other.variable {  }
-.invalid {  }
-.markup.quote { color: rgba(133, 153, 0, 1); }
-.markup.list { color: rgba(181, 137, 0, 1); }
-.markup.bold { color: rgba(211, 54, 130, 1); }
-.markup.italic { color: rgba(211, 54, 130, 1); }
-.markup.inline.raw { color: rgba(42, 161, 152, 1); }
-.markup.heading { color: rgba(38, 139, 210, 1); }
-.markup.heading.setext { color: rgba(38, 139, 210, 1); }
